Плагин: - обновлены модели компонентов com_content, com_contact, com_tags из версии Joomla 4.3.3. На Joomla 5.0.0-alpha2 работают. - модели перенесены в src\Models, чтоб весь плагин был в одной куче - удален неиспользуемый код
